requirements.txt
Flask-script
PyMongo
Flask-PyMongo
requests
Flask-WTF
flask-login
Flask-Security
Flask-MongoEngine
